+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `uploads_pause_on_fail` is set to `true`.
 - Emit the `success` event for successfully uploaded files if users cancel the
   upload mid-way through uploading a set of files in the Computer view.
+- Improve UI when the `multiselect` option is configured to `false`.
 
 ## [2.2.4] - 2020-04-20